Started using use-package for lazy loading
I started using use-package for lazy loading. I also made many many
changes. No time to explain at the moment. Lots more working needs to
be done and it needs to be re-organized. This is just a WIP that is
not complete, but I need to check this in now for use elsewhere.